move images to parent folder for better structure and remove unused original images
