+++ src/routes/about/assets/about.css
@@ -90,7 +90,7 @@
   padding: 4rem;
   padding-right: 10rem;
   z-index: 98;
-  background-image: url("../../assets/images/about-us/section-one.jpeg");
+  background-image: url("./section-one.jpeg");
 }
 .about__company__content h2,
 .about__company__content h3,
@@ -123,7 +123,7 @@
 }
 
 .about__dedication {
-  background-image: url("../../assets/images/about-us/section-two.jpeg");
+  background-image: url("./section-two.jpeg");
 }
 
 .about__dedication__content {
@@ -168,7 +168,7 @@
   width: 100%;
   height: 20rem;
   bottom: 0;
-  background-image: url("../../assets/images/about-us/section-three.jpeg");
+  background-image: url("./section-three.jpeg");
   background-position: center;
   background-repeat: repeat;
   background-size: cover;
