+++ src/routes/about/assets/section-one.jpeg
