+++ src/routes/about/components/AboutCustomers.jsx
@@ -1,6 +1,6 @@
-import ImgOne from "../../../assets/images/about-us/img-1.jpeg";
-import ImgTwo from "../../../assets/images/about-us/img-2.jpeg";
-import ImgThree from "../../../assets/images/about-us/img-3.jpeg";
+import ImgOne from "../assets/img-1.jpeg";
+import ImgTwo from "../assets/img-2.jpeg";
+import ImgThree from "../assets/img-3.jpeg";
 
 const AboutCustomers = () => {
   return (
